feat: add importation on Meteo, modifications on Location and suppression of Demography
feat: add importation on Meteo, modifications on Location and suppression of Demography

feat: add importation on Meteo

# Conflicts:
#	src/main/java/fr/diginamic/echolink/infrastructure/meteo/in/mapper/MeteoQueryMapper.java

+++ src/main/java/fr/diginamic/echolink/infrastructure/location/in/dto/LocationQuery.java
@@ -3,6 +3,6 @@ package fr.diginamic.echolink.infrastructure.location.in.dto;
 public record LocationQuery (
     String name,
     String postalCode,
-    float longitude,
-    float latitude) {
+    double longitude,
+    double latitude) {
 }
